Updated file structure + re-wrote test
