+++ node/src/commands/genesis/input.rs
@@ -1,0 +1,46 @@
+use miden_objects::accounts::AccountType;
+use serde::Deserialize;
+
+// INPUT HELPER STRUCTS
+// ================================================================================================
+
+/// *Input types are helper structures designed for parsing and deserializing configuration files.
+/// They serve as intermediary representations, facilitating the conversion from
+/// placeholder types (like `GenesisInput`) to internal types (like `GenesisState`).
+#[derive(Debug, Deserialize)]
+pub struct GenesisInput {
+    pub version: u64,
+    pub timestamp: u64,
+    pub accounts: Vec<AccountInput>,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(tag = "type")]
+pub enum AccountInput {
+    BasicWallet(BasicWalletInputs),
+    BasicFungibleFaucet(BasicFungibleFaucetInputs),
+}
+
+#[derive(Debug, Deserialize)]
+pub struct BasicWalletInputs {
+    pub mode: AccountType,
+    pub init_seed: String,
+    pub auth_scheme: AuthSchemeInput,
+    pub auth_seed: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct BasicFungibleFaucetInputs {
+    pub mode: AccountType,
+    pub init_seed: String,
+    pub auth_scheme: AuthSchemeInput,
+    pub auth_seed: String,
+    pub token_symbol: String,
+    pub decimals: u8,
+    pub max_supply: u64,
+}
+
+#[derive(Debug, Deserialize)]
+pub enum AuthSchemeInput {
+    RpoFalcon512,
+}
